requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.1
openai>=1.0.0
python-dotenv>=0.19.0
elevenlabs>=0.2.0
vosk>=0.3.45
sounddevice>=0.4.5
numpy>=1.20.0 